net/nanocoap: make nanocoap_get_blockwise_url_buf return filesize

+++ sys/include/net/nanocoap_sock.h
@@ -77,7 +77,7 @@ ssize_t nanocoap_get_blockwise(sock_udp_ep_t *remote, const char *path,
 int nanocoap_get_blockwise_url(const char *url, coap_blksize_t blksize,
                                coap_blockwise_cb_t callback, void *arg);
 
-int nanocoap_get_blockwise_url_buf(const char *url, coap_blksize_t blksize,
+ssize_t nanocoap_get_blockwise_url_buf(const char *url, coap_blksize_t blksize,
                                    uint8_t *buf, size_t len);
 /**
  * @brief   Simple synchronous CoAP request
